# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My vinyls collection",
        "description": "My vinyls collection Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter your bearer token in the format **Bearer &lt;token>**"
          }
    },
    "tags": [
        {
            "name": "Vinyl",
            "description": "API for users in the system"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/sign-up": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Adding one user",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "The user we want to create",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "repeat_password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "409": {
                        "description": "This username is already in use !"
                    },
                    "400": {
                        "description": "Something went wrong with the database operation"
                    },
                    "401": {
                        "description": "Username or password is incorrect!"
                    },
                    "500": {
                        "description": "Username available but something is wrong !"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login an User and return informations",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "The user we want to login",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logged in"
                    },
                    "400": {
                        "description": "Error with database"
                    },
                    "401": {
                        "description": "Username or password is incorrect !"
                    }
                }
            }
        },
        "/vinyl/get/": {
            "get": {
                "security": [{ "Bearer": [] }],
                "tags": [
                    "Vinyl"
                ],
                "summary": "Get all vinyls from the user with his id in JWT token informations",
                "responses": {
                    "200": {
                        "description": "All vinyls from the user"
                    },
                    "400": {
                        "description": "Error in database operation"
                    }
                }
            }
        },
        "/vinyl/create/": {
            "post": {
                "security": [{ "Bearer": [] }],
                "tags": [
                    "Vinyl"
                ],
                "summary": "Adding one vinyl from the user with his id in JWT token informations",
                "parameters": [
                    {
                        "name": "vinyl",
                        "in": "body",
                        "description": "Vinyl we want to create",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "artist": {
                                    "type": "string"
                                },
                                "album": {
                                    "type": "string"
                                },
                                "label": {
                                    "type": "string"
                                },
                                "release_date": {
                                    "type": "string",
                                    "format": "date"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New vinyl create successfully",
                        "schema": {
                            "$ref": "#/definitions/Vinyl"
                        }
                    },
                    "400": {
                        "description": "Error in database operation"
                    }
                }
            }
        },
        "/vinyl/update/{id}": {
            "put": {
                "security": [{ "Bearer": [] }],
                "tags": [
                    "Vinyl"
                ],
                "summary": "Update one vinyl with this id in path & User id in JWT token informations",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vinyl we want to update",
                        "schema": {
                            "id": {
                                "type": "number",
                                "uniqueItems": true
                            }
                        }
                    },
                    {
                        "name": "vinyl",
                        "in": "body",
                        "description": "Vinyl we want to update",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "artist": {
                                    "type": "string"
                                },
                                "album": {
                                    "type": "string"
                                },
                                "label": {
                                    "type": "string"
                                },
                                "release_date": {
                                    "type": "string",
                                    "format": "date"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vinyl updated successfully",
                        "schema": {
                            "$ref": "#/definitions/Vinyl"
                        }
                    },
                    "400": {
                        "description": "Error in database operation"
                    }
                }
            }
        },
        "/vinyl/delete/{id}": {
            "delete": {
                "security": [{ "Bearer": [] }],
                "tags": [
                    "Vinyl"
                ],
                "summary": "Delete one vinyl with this id in the path & User id in JWT token informations",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vinyl we want to delete",
                        "schema": {
                            "$ref": "#/definitions/Vinyl"
                        }
                    }
                ],
                "schema": {
                    "type": "object",
                    "properties": {
                        "fk_id_user": {
                            "type": "int"
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vinyl deleted successfully"
                    },
                    "400": {
                        "description": "Error in database operation"
                    }
                }
            }
        }
    },
    "definitions": {
        "Vinyl": {
            "required": [
                "id",
                "fk_id_user",
                "artist",
                "album",
                "label",
                "release_date"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "fk_id_user": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "artist": {
                    "type": "string"
                },
                "album": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string",
                    "format": "date"
                },
                "date_added": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
        "Users": {
            "required": [
                "id",
                "username",
                "secret_password",
                "registered",
                "last_login"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "username": {
                    "type": "string",
                    "uniqueItems": true
                },
                "secret_password": {
                    "type": "string"
                },
                "registered": {
                    "type": "string",
                    "format": "date"
                },
                "last_login": {
                    "type": "string",
                    "format": "date"
                }
            }
        }
    }
}